+++ examples/WireMock.Net.Console.Net452.Classic/MainApp.cs
@@ -13,6 +13,7 @@ using WireMock.Server;
 using WireMock.Settings;
 using WireMock.Util;
 using System.Threading.Tasks;
+using WireMock.Types;
 
 namespace WireMock.Net.ConsoleApplication
 {
@@ -346,7 +347,7 @@ namespace WireMock.Net.ConsoleApplication
                     .WithHeader("Transformed-Postman-Token", "token is {{request.headers.Postman-Token}}")
                     .WithHeader("xyz_{{request.headers.Postman-Token}}", "token is {{request.headers.Postman-Token}}")
                     .WithBody(@"{""msg"": ""Hello world CATCH-ALL on /*, {{request.path}}, add={{Math.Add request.query.start.[0] 42}} bykey={{request.query.start}}, bykey={{request.query.stop}}, byidx0={{request.query.stop.[0]}}, byidx1={{request.query.stop.[1]}}"" }")
-                    .WithTransformer()
+                    .WithTransformer(TransformerType.Handlebars)
                     .WithDelay(TimeSpan.FromMilliseconds(100))
                 );
