+++ src/WireMock.Net/ResponseBuilders/ITransformResponseBuilder.cs
@@ -1,4 +1,6 @@
-﻿namespace WireMock.ResponseBuilders
+﻿using WireMock.Types;
+
+namespace WireMock.ResponseBuilders
 {
     /// <summary>
     /// The TransformResponseBuilder interface.
@@ -6,11 +8,19 @@
     public interface ITransformResponseBuilder : IDelayResponseBuilder
     {
         /// <summary>
-        /// The with transformer.
+        /// Use the Handlebars.Net ResponseMessage transformer.
         /// </summary>
         /// <returns>
         /// The <see cref="IResponseBuilder"/>.
         /// </returns>
         IResponseBuilder WithTransformer(bool transformContentFromBodyAsFile = false);
+
+        /// <summary>
+        /// Use a specific ResponseMessage transformer.
+        /// </summary>
+        /// <returns>
+        /// The <see cref="IResponseBuilder"/>.
+        /// </returns>
+        IResponseBuilder WithTransformer(TransformerType transformerType, bool transformContentFromBodyAsFile = false);
     }
 }
