+++ src/WireMock.Net/Transformers/IResponseMessageTransformer.cs
@@ -1,0 +1,7 @@
+﻿namespace WireMock.Transformers
+{
+    interface IResponseMessageTransformer
+    {
+        ResponseMessage Transform(RequestMessage requestMessage, ResponseMessage original, bool useTransformerForBodyAsFile);
+    }
+}
